+++ changelog.md
@@ -4,6 +4,7 @@ v0.0.20-alpha
 * Add recipe book functionality to the concocter
 * Add Essence Transmuter, to convert one source of essence into another
 * Add Mystical Relics, fragments of which drop from certain mobs, to study for theories
+* Increment to Forge version 37.0.97
 
 v0.0.19-alpha
 -------------